completely rework directive handling
Finish rework of Directive scanning. Was able to clean up the parsing so
we don't rely on a Python string (yay!) and so was able to get rid of
some ugly methods I hated in the ScannerIterator (double yay!).
Directives unfortunately need two passes to handle but that seems a
small price to pay.