start working on rules
In order to better understand how to handle conditional blocks, I think
I need to start working on handling Rules.

I've also added a check in the top level execute() for an eval() coming
back non-empty (which is where I ran aground on the rocks of conditional
blocks).

Very much a starting point work in progress!