get_vline() now decides to build a RecipeVirtualLine on <tab>
instead of pushing the decision of RecipeVirtualLine vs VirtualLine up
higher in the stack, let's make that decision here.  If we've made the
wrong decision (such as <tab>ifdef before first recipe) then we'll
convert the line later.

This is a small step toward better handling of Rules and Recipes