huge fix in handling nested blocks
A nested block is the contents of an Conditional Block. A Conditional
Block can contain other Conditional Blocks or makefile code stored in a
LineBlock.

The contents of a Block (ConditionalBlock or LineBlock) can't be parsed
until the condition is known. At execute(), the Block is eval'd which
will parse the Block itself. The result will then also need to be
eval'd.

A Block can contain Block(s) can contain Block(s).  It's turtles all the
way down.  In order to handle such nested constructs, I've moved the
statement execute into a function which can be called recursively. The
recusive execute function will update the SymbolTable, a RulesDB, and a
list of current Rules. (The current Rules will be updated when a Recipe
is found in the nested Block(s)

When an eval() returns a list of statements, the execute function will
call itself recursively, updating the same symtable, rules, etc, data
structures. At the end of the statement execute, we will have a fully
populated SymbolTable and RulesDB.