tinkering with sub-makes
Idea is to launch my own wrapper from the shell. We need to let the
shell interpret the command line, then we'll capture the arguments.

The sub-make capture will instantiate an entire new token-z-parser and
symbol table. We will run the sub-make in the same process, blocking the
parent make.

We want to be a debugger. Everything needs to be in the same process
space so we can access everything simply (no shared memory
shenanigans).

This is very experimental and has been blocking me for quite a while.