add submake $(MAKE) support
Submake support is through a brute force design of simply running a
helper script that prints the sub-make's args to stdout. (We need the
shell to interpret the command line.)

The shell.py code checks for a command starting with py-submake (the
name of the helper script). If found, the return of the sub-process will
have an 'is_submake' flag set. The calling code will check that flag
and, if set, run a new tokenzparser+execute on that makefile.

Needs more testing.