fix nested ifdef blocks
My original solution to handling nested ifdef blocks was incorrect.
I would .eval the elements inside (LineBlock or another conditional).
That .eval of a LineBlock just parsed the lines. The .eval of the
conditional would evaluate the conditional.

ifneq (0,1)
foo:=1
ifneq ($(foo),1)
    $(error should not hit this)
endif
endif

The .eval of the conditional would be evaluated before the assignment
and so 'foo' would be empty.