add "layers" to symbol table
In order to support rule-specific variables, let's add a 'layer' scheme
to the symbol table.  A "layer" is another dict with key:value as
name:Entry. The layers are stored in an array, with [0] as the topmost
layer.

Variable lookups search starting at [0]. Variables only added to layer
[0].

Removed the old push/pop variable code.

Need to add test code.